Update ReserveElockerServlet.java
Implementing the controller method, doPost, to response locker reservation requests. Since LockerReservationFacade is a stateful session bean, as the first step, it checks if the bean is not yet in the HTTP session (the first client request for a locker reservation) and if not a new LockerReservationFacade instance is added to the session object. Then the business methods are called on that instance to process the locker reservation.